+++ android/apollo/src/com/andrew/apollo/MusicPlaybackService.java
@@ -51,7 +51,7 @@ import android.support.v4.content.ContextCompat;
 import com.andrew.apollo.cache.ImageCache;
 import com.andrew.apollo.cache.ImageFetcher;
 import com.andrew.apollo.provider.FavoritesStore;
-import com.andrew.apollo.provider.RecentSongStore;
+import com.andrew.apollo.provider.RecentStore;
 import com.andrew.apollo.ui.activities.AudioPlayerActivity;
 import com.andrew.apollo.utils.MusicUtils;
 import com.frostwire.android.BuildConfig;
@@ -482,7 +482,7 @@ public class MusicPlaybackService extends Service {
      */
 //    private RecentStore mRecentsCache;
 
-    private RecentSongStore mRecentsCache;
+    private RecentStore mRecentsCache;
     /**
      * Favorites database
      */
@@ -569,8 +569,7 @@ public class MusicPlaybackService extends Service {
 
     private void initService() {
         // Initialize the favorites and recents databases
-//        mRecentsCache = RecentStore.getInstance(this);
-        mRecentsCache = RecentSongStore.getInstance(this);
+        mRecentsCache = RecentStore.getInstance(this);
         mFavoritesCache = FavoritesStore.getInstance(this);
 
         // Initialize the notification helper
@@ -1677,16 +1676,8 @@ public class MusicPlaybackService extends Service {
             }
             MusicPlaybackService musicPlaybackService = musicPlaybackServiceWeakReference.get();
             if (musicPlaybackService.mRecentsCache == null) {
-//                musicPlaybackService.mRecentsCache = RecentStore.getInstance(musicPlaybackService);
-                musicPlaybackService.mRecentsCache = RecentSongStore.getInstance(musicPlaybackService);
+                musicPlaybackService.mRecentsCache = RecentStore.getInstance(musicPlaybackService);
             }
-
-//            long albumId = musicPlaybackService.getAlbumId();
-//            String albumName = musicPlaybackService.getAlbumName();
-//            String artistName = musicPlaybackService.getArtistName();
-//            String songCount =  MusicUtils.getSongCountForAlbum(musicPlaybackService, albumId);
-//            String releaseDateForAlbum = MusicUtils.getReleaseDateForAlbum(musicPlaybackService, albumId);
-//            musicPlaybackService.mRecentsCache.addAlbumId(albumId, albumName, artistName, songCount, releaseDateForAlbum);
 
             long songId = musicPlaybackService.getAudioId();
             String songName = musicPlaybackService.getTrackName();
